Rename UseCasePage and add UseCase SubComponent
